Move finetune_lora_config.py to config/ directory.
