+++ src/invoke_training/training/config/finetune_lora_config.py
@@ -129,6 +129,12 @@ class FinetuneLoRAConfig(BaseModel):
     # applied.
     cache_text_encoder_outputs: bool = False
 
+    # If True, the VAE will be applied to all of the images in the dataset before starting training and the results will
+    # be cached to disk. This reduces the VRAM requirements during training (don't have to keep the VAE in VRAM), and
+    # speeds up training (don't have to run the VAE encoding step). This option can only be enabled if all
+    # non-deterministic image augmentations are disabled (i.e. center_crop=True, random_flip=False).
+    cache_vae_outputs: bool = False
+
     # If True, models will be kept in CPU memory and loaded into GPU memory one-by-one while generating validation
     # images. This reduces VRAM requirements at the cost of slower generation of validation images.
     enable_cpu_offload_during_validation: bool = False
