+++ src/invoke_training/training/finetune_lora/finetune_lora_sd.py
@@ -25,9 +25,7 @@ from invoke_training.lora.injection.stable_diffusion import (
     inject_lora_into_clip_text_encoder,
     inject_lora_into_unet,
 )
-from invoke_training.training.finetune_lora.finetune_lora_config import (
-    FinetuneLoRAConfig,
-)
+from invoke_training.training.config.finetune_lora_config import FinetuneLoRAConfig
 from invoke_training.training.shared.accelerator_utils import (
     get_mixed_precision_dtype,
     initialize_accelerator,
