+++ src/invoke_training/training/shared/data/data_loaders/image_caption_sdxl_dataloader.py
@@ -4,7 +4,7 @@ import torch
 from torch.utils.data import DataLoader
 from transformers import PreTrainedTokenizer
 
-from invoke_training.training.finetune_lora.finetune_lora_config import DatasetConfig
+from invoke_training.training.config.finetune_lora_config import DatasetConfig
 from invoke_training.training.shared.data.datasets.hf_dir_image_caption_dataset import (
     HFDirImageCaptionDataset,
 )
